fixing problems with std::abs

+++ TRExFitter/CorrelationMatrix.h
@@ -20,7 +20,7 @@ public:
     //
     void AddNuisPar(const std::string& p);
     void SetCorrelation(const std::string& p0, const std::string& p1,float corr);
-    float GetCorrelation(const std::string& p0, const std::string& p1);
+    double GetCorrelation(const std::string& p0, const std::string& p1);
     void Draw(const std::string& path, const double corrMin = -1.);
 
     //
